Refactor RESTORE_GRAPH_STATE
Instead of setting `repositioning` to `true` in the RESTORE_GRAPH_STATE
branch, I've decided to delegate that responsibility. It has nothing to
do with restoring, and it should be handled separately.

Sadly, I can't do the same thing (easily) with the `selected` prop.
I added a comment to explain why it's necessary, and will need to
leave it at that.